refactor: Hilt Module 구성 이름 변경
