+++ data/src/main/java/com/github/dodobest/data/di/APIModule.kt
@@ -10,7 +10,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal object UpbitAPIModule {
+internal object APIModule {
     @Provides
     @Singleton
     fun provideUpbitAPI(
