+++ data/src/main/java/com/github/dodobest/data/di/NetworkModule.kt
@@ -6,6 +6,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import hu.akarnokd.rxjava3.retrofit.RxJava3CallAdapterFactory
 import okhttp3.OkHttpClient
+import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.util.concurrent.TimeUnit
 import javax.inject.Singleton
@@ -17,6 +18,7 @@ internal object NetworkModule {
     private const val CONNECT_TIMEOUT = 3L
     private const val READ_TIMEOUT = 3L
     private const val WRITE_TIMEOUT = 3L
+    private const val UPBIT_BASE_URL = "https://api.upbit.com/v1/"
 
     @Provides
     @Singleton
@@ -38,5 +40,20 @@ internal object NetworkModule {
     @Singleton
     fun provideRxJava3CallAdapter(): RxJava3CallAdapterFactory {
         return RxJava3CallAdapterFactory.create()
+    }
+
+    @Provides
+    @Singleton
+    fun provideUpbitRetrofit(
+        rxJava3CallAdapterFactory: RxJava3CallAdapterFactory,
+        gsonConverterFactory: GsonConverterFactory,
+        okHttpClient: OkHttpClient,
+    ): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(UPBIT_BASE_URL)
+            .addCallAdapterFactory(rxJava3CallAdapterFactory)
+            .addConverterFactory(gsonConverterFactory)
+            .client(okHttpClient)
+            .build()
     }
 }
