+++ data/src/main/java/com/github/dodobest/data/di/RepositoryModule.kt
@@ -10,7 +10,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal interface UpbitRepositoryModule {
+internal interface RepositoryModule {
     @Binds
     @Singleton
     fun provideUpbitRepository(
