--- data/src/main/java/com/github/dodobest/data/di/UpbitRetrofitModule.kt
@@ -1,32 +1,0 @@
-package com.github.dodobest.data.di
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import hu.akarnokd.rxjava3.retrofit.RxJava3CallAdapterFactory
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal object UpbitRetrofitModule {
-    private const val UPBIT_BASE_URL = "https://api.upbit.com/v1/"
-
-    @Provides
-    @Singleton
-    fun provideUpbitRetrofit(
-        rxJava3CallAdapterFactory: RxJava3CallAdapterFactory,
-        gsonConverterFactory: GsonConverterFactory,
-        okHttpClient: OkHttpClient,
-    ): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(UPBIT_BASE_URL)
-            .addCallAdapterFactory(rxJava3CallAdapterFactory)
-            .addConverterFactory(gsonConverterFactory)
-            .client(okHttpClient)
-            .build()
-    }
-}
